move single-person fetching to separate file
