+++ lib/contentful/index.js
@@ -7,6 +7,7 @@ export * from './collaborations'
 // export * from './organizations'
 export * from './groups'
 export * from './people'
+export * from './person'
 export * from './projects'
 export * from './teams'
